requirements.txt
# ∫Û∂À“¿¿µ
flask
python-docx
pdfplumber
pytesseract
pillow
opencv-python
# «∞∂À“¿¿µ
